[pkg/datadog] Expose TranslatorFromConfig
