+++ exporter/datadogexporter/metrics_exporter.go
@@ -18,7 +18,6 @@ import (
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes/source"
 	otlpmetrics "github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/metrics"
-	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
@@ -30,6 +29,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/metrics"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/metrics/sketches"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/scrub"
+	datadogconfig "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/datadog/config"
 )
 
 type metricsExporter struct {
@@ -50,48 +50,6 @@ type metricsExporter struct {
 	getPushTime func() uint64
 }
 
-// translatorFromConfig creates a new metrics translator from the exporter
-func translatorFromConfig(set component.TelemetrySettings, cfg *Config, attrsTranslator *attributes.Translator, sourceProvider source.Provider, statsOut chan []byte) (*otlpmetrics.Translator, error) {
-	options := []otlpmetrics.TranslatorOption{
-		otlpmetrics.WithDeltaTTL(cfg.Metrics.DeltaTTL),
-		otlpmetrics.WithFallbackSourceProvider(sourceProvider),
-	}
-
-	if isMetricRemappingDisabled() {
-		set.Logger.Warn("Metric remapping is disabled in the Datadog exporter. OpenTelemetry metrics must be mapped to Datadog semantics before metrics are exported to Datadog (ex: via a processor).")
-	} else {
-		options = append(options, otlpmetrics.WithRemapping())
-	}
-
-	if cfg.Metrics.HistConfig.SendAggregations {
-		options = append(options, otlpmetrics.WithHistogramAggregations())
-	}
-
-	if cfg.Metrics.SummaryConfig.Mode == SummaryModeGauges {
-		options = append(options, otlpmetrics.WithQuantiles())
-	}
-
-	if cfg.Metrics.ExporterConfig.InstrumentationScopeMetadataAsTags {
-		options = append(options, otlpmetrics.WithInstrumentationScopeMetadataAsTags())
-	}
-
-	options = append(options, otlpmetrics.WithHistogramMode(otlpmetrics.HistogramMode(cfg.Metrics.HistConfig.Mode)))
-
-	var numberMode otlpmetrics.NumberMode
-	switch cfg.Metrics.SumConfig.CumulativeMonotonicMode {
-	case CumulativeMonotonicSumModeRawValue:
-		numberMode = otlpmetrics.NumberModeRawValue
-	case CumulativeMonotonicSumModeToDelta:
-		numberMode = otlpmetrics.NumberModeCumulativeToDelta
-	}
-	options = append(options, otlpmetrics.WithNumberMode(numberMode))
-	options = append(options, otlpmetrics.WithInitialCumulMonoValueMode(
-		otlpmetrics.InitialCumulMonoValueMode(cfg.Metrics.SumConfig.InitialCumulativeMonotonicMode)))
-
-	options = append(options, otlpmetrics.WithStatsOut(statsOut))
-	return otlpmetrics.NewTranslator(set, attrsTranslator, options...)
-}
-
 func newMetricsExporter(
 	ctx context.Context,
 	params exporter.Settings,
@@ -103,7 +61,7 @@ func newMetricsExporter(
 	metadataReporter *inframetadata.Reporter,
 	statsOut chan []byte,
 ) (*metricsExporter, error) {
-	tr, err := translatorFromConfig(params.TelemetrySettings, cfg, attrsTranslator, sourceProvider, statsOut)
+	tr, err := datadogconfig.TranslatorFromConfig(params.TelemetrySettings, cfg.Metrics, attrsTranslator, sourceProvider, statsOut)
 	if err != nil {
 		return nil, err
 	}
